test(fixtures): centralize in-process adapter setup
